requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.0
flask==3.0.0
